auth/requirements.txt
flask
flask_pymongo
flask_session
ldap3
redis
pyodbc